% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unsup.R
\name{augment_data}
\alias{augment_data}
\title{augment data by a univariate boostrap of the features}
\usage{
augment_data(data)
}
\arguments{
\item{data}{a matrix or data.frame}
}
\value{
a data.frame with row dimension twice that of data, with an additional factor column
indicating whether the data is from the input, data (labelled "observed"), or was generated by
the univariate bootstrap (labelled "synthetic")
}
\description{
augment data by a univariate boostrap of the features
}

