% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/edarf.R
\docType{package}
\name{edarf}
\alias{edarf}
\alias{edarf-package}
\title{edarf: Exploratory Data Analysis Using Random Forests}
\description{
This package provides utilities for exploratory data analysis (EDA) with random forests. It allows the calculation and visualization of partial dependence for single and multiple predictors, interactiondetection and proximity of observations.
}
\section{Partial Dependence}{

}

\section{Interaction Detection}{

}

\section{Proximity}{

}

