% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prox.R
\name{extract_proximity.RandomForest}
\alias{extract_proximity.RandomForest}
\title{Extract proximity matrix from objects returned from \code{cforest}}
\usage{
\method{extract_proximity}{RandomForest}(fit, newdata = NULL, ...)
}
\arguments{
\item{fit}{object of class 'RandomForest' from \code{cforest}}

\item{newdata}{a data.frame with the same columns as the training data}

\item{...}{additional arguments to pass to proximity}
}
\value{
an n by n matrix
}
\description{
Extract proximity matrix from objects returned from \code{cforest}
}

