% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pd.R
\name{fix_classes}
\alias{fix_classes}
\title{Matches column classes of the input data frame to the output}
\usage{
fix_classes(var, df, pred)
}
\arguments{
\item{var}{character vector of column names to match}

\item{df}{imput dataframe}

\item{pred}{output dataframe}
}
\value{
dataframe \code{pred} with \code{var} column classes matched to those in \code{df}
}
\description{
Matches column classes of the input data frame to the output
}

