% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ci.R
\name{inf_jackknife}
\alias{inf_jackknife}
\title{Bias corrected infinitesimal jackknife variance estimator for predictions given a matrix of tree predictions}
\usage{
inf_jackknife(pred, B, N)
}
\arguments{
\item{pred}{matrix with n rows and B columns, where n is the number of observations, and B the number of trees in the forest}

\item{B}{number of trees}

\item{N}{matrix with n rows and B columns, where each entry gives the number of times observation i appears in tree j}
}
\value{
a vector of variance estimates for each observation
}
\description{
Essentially code from randomForestCI (\url{https://github.com/swager/randomForestCI}) with some modifications, and factored out so that it can be made more generic
}
\examples{
\dontrun{
library(randomForest)
data(swiss)

fit <- randomForest(Fertility ~ ., swiss, keep.inbag = TRUE)
pred <- predict(fit, newdata = swiss, predict.all = TRUE)
inf_jackknife(pred$individual, fit$ntree, fit$inbag)
}
}

