% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int.R
\name{interaction_importance}
\alias{interaction_importance}
\title{Find interaction_importances by comparing marginal/joint variable importance}
\usage{
interaction_importance(fit, var, nperm = 100, parallel = FALSE, ...)
}
\arguments{
\item{fit}{object of class 'RandomForest', 'randomForest', or 'rfsrc'}

\item{var}{a character vector of the predictors of interest, which must match the input matrix}

\item{nperm}{integer, number of permutations to run, default is 100}

\item{parallel}{logical, whether to use a registered parallel backend}

\item{...}{additional arguments to be passed}
}
\value{
a named vector
}
\description{
Computes the individual importance of a set of variables and compares it to the importance of the variables together
}
\examples{
\dontrun{
library(randomForest)
data(iris)

fit <- randomForest(Species ~ ., iris)
interaction_importance(fit, c("Petal.Width", "Petal.Length"), 10, data = iris)
}
}

