% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ivar_points}
\alias{ivar_points}
\title{Creates a prediction vector for variables to decrease computation time}
\usage{
ivar_points(df, x, cutoff = 10)
}
\arguments{
\item{df}{the dataframe used to fit the random forest, extracted from the fitted object}

\item{x}{a character vector of length 1 indicating the variable in `df` to be extracted}

\item{cutoff}{an integer indicating the maximal length of the vector to be used for prediction}
}
\value{
a vector of unique values taken by `x` of length < `cutoff`
}
\description{
Creates a prediction vector for variables to decrease computation time
}

