% Generated by roxygen2 (4.0.1): do not edit by hand
\name{partial_dependence}
\alias{partial_dependence}
\title{Partial dependence using random forests}
\usage{
partial_dependence(fit, df, var, surv = FALSE, cores = 1, ...)
}
\arguments{
\item{fit}{an object of class 'RandomForest-class' returned from \code{cforest}, or an object
of class 'randomForest' returned from \code{randomForest}}

\item{df}{the dataframe used to fit the model, if the model is a party object of class 'RandomForest'
this option can be omitted and the dataframe will be extracted from the object}

\item{var}{a character vector of the predictors of interest, which must match the input
matrix in the call to \code{cforest}, \code{randomForest}, or \code{randomForestSRC}}

\item{surv}{logical, indicates whether or not the response is right-censored}

\item{cores}{indicates the number of cores to use. parallelization occurs in the prediction
on the grid of possible values taken by all combinations of `var`}

\item{...}{arguments to be passed to 'ivar_points'}
}
\value{
a dataframe with columns for each predictor in `var` and the fitted value for
each set of values taken by the values of 'var' averaged within the values of predictors
in the model but not in `var`
}
\description{
Calculates the partial dependence of the response on an arbitrary dimensional set of predictors
from a fitted random forest object from the Party, randomForest, or randomForestSRC packages
}
\examples{
require(randomForest)
require(party)
require(randomForestSRC)
require(parallel)
data(iris)
fit_rf <- randomForest(Species ~ ., iris)
fit_pt <- cforest(Species ~ ., iris, controls = cforest_control(mtry = 2))
fit_rfsrc <- rfsrc(Species ~ ., iris)

pd_rf <- partial_dependence(fit_rf, iris, "Petal.Width", detectCores())
pd_pt <- partial_dependence(fit_pt, iris, "Petal.Width", detectCores())
pd_rfsrc <- partial_dependence(fit_rfsrc, iris, "Petal.Width", detectCores())

pd_int_rf <- partial_dependence(fit_rf, iris, c("Petal.Width", "Sepal.Length"), detectCores())
pd_int_pt <- partial_dependence(fit_pt, iris, c("Petal.Width", "Sepal.Length"), detectCores())
pd_int_rfsrc <- partial_dependence(fit_rfsrc, iris, c("Petal.Width", "Sepal.Length"), detectCores())
}
\author{
Zachary M. Jones, \email{zmj@zmjones.com}
}

