% Generated by roxygen2 (4.0.1): do not edit by hand
\name{party_partial_dependence}
\alias{party_partial_dependence}
\title{Partial dependence using Party random forests}
\usage{
party_partial_dependence(fit, var, surv = FALSE, cores = 1, ...)
}
\arguments{
\item{fit}{an object of class 'RandomForest-class' returned from \code{cforest}}

\item{var}{a character vector of the predictors of interest, which must match the input
matrix in the call to \code{cforest}}

\item{surv}{logical, indicates whether or not the response is right-censored}

\item{cores}{indicates the number of cores to use. parallelization occurs in the prediction
on the grid of possible values taken by all combinations of `var`}

\item{...}{arguments to be passed to 'ivar_points'}
}
\value{
a dataframe with columns for each predictor in `var` and the fitted value for
each set of values taken by the values of 'var' averaged within the values of predictors
in the model but not in `var`
}
\description{
Calculates the partial dependence of the response on an arbitrary dimensional set of predictors
from a fitted random forest object from the Party package
}
\examples{
data(iris)

fit <- party::cforest(Species ~ ., data = iris)
pd <- party_partial_dependence(fit, "Petal.Width", parallel::detectCores())
pd_int <- party_partial_dependence(fit, c("Petal.Width", "Sepal.Length"), parallel::detectCores())
}
\author{
Zachary M. Jones, \email{zmj@zmjones.com}
}

