% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pd_inner}
\alias{pd_inner}
\title{Calculates the partial dependence of 'var' on the response for a particular value of 'var'}
\usage{
pd_inner(fit, df, var, rng, type, i)
}
\arguments{
\item{fit}{an object of class 'RandomForest-class' returned from \code{cforest}, an object
of class 'randomForest' returned from \code{randomForest}, or an object of class 'rfsrc'
returned from \code{rfsrc}}

\item{df}{the dataframe used to fit the model, if the model is a party object of class 'RandomForest'
this option can be omitted and the dataframe will be extracted from the object}

\item{var}{a character vector of the predictors of interest, which must match the input
matrix in the call to \code{cforest}, \code{randomForest}, or \code{randomForestSRC}}

\item{rng}{a dataframe with columns equal to the number of variables in 'var' and rows less
than or equal to 'cutoff' as defined in the call to 'ivar_points'}

\item{type}{the class of the dependent variable, generated by 'partial_dependence.'
can equal "numeric", "factor," or "survival"}

\item{i}{the index position in 'rng' for which the variables in 'var' are set}
}
\value{
a vector with the values the variables specified by 'var' are set at and the summary fitted value
}
\description{
Calculates the fitted values of observations with one or more variables set to a particular value
and returns a summary of these fitted values: means for regression and survival analysis and
the modal category for classification,
}

