% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_int}
\alias{plot_int}
\title{Plot interaction detection using variable importance from random forests}
\usage{
plot_int(int, labels = NULL, vertical_labels = FALSE, xlab = "feature",
  ylab = "importance", title = NULL)
}
\arguments{
\item{int}{object of class "interaction_importance", a named numeric vector}

\item{labels}{optional character vector of \code{length(int)} replacing \code{names(int)}}

\item{vertical_labels}{optional logical rotating labels vertically}

\item{xlab}{optional x-axis label, default "feature"}

\item{ylab}{optional y-axis label, default "importance"}

\item{title}{optional title, default empty string}
}
\value{
a ggplot2 object
}
\description{
Plot interaction detection using variable importance from random forests
}
\examples{
\dontrun{

library(party)
data(iris)
fit <- cforest(Species ~ ., iris, controls = cforest_unbiased(mtry = 2))
int <- interaction_importance(fit, c("Petal.Width", "Petal.Length"), 10)
plot_int(int)
}
}

