% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_twoway_partial}
\alias{plot_twoway_partial}
\title{Plot two-way partial dependence from random forests}
\usage{
plot_twoway_partial(var, pred, var_lab, grid, smooth = FALSE, ylab = NULL,
  xlab = NULL, title = NULL)
}
\arguments{
\item{var}{numeric, integer, or factor values the variable takes on.}

\item{pred}{numeric, factor, or integer vector of model predictions
results in a scatterplot for integer or numeric, and a bar plot for factors}

\item{var_lab}{character or factor that specifies the variables' labels
if multiple variables are in 'var'}

\item{grid}{integer vector of length 2, which specifies the number of rows
and columns in the plot grid if 'var_lab' is specified}

\item{smooth}{logical, whether or not a GAM smoother should be overlaid, only possible if
'var' and 'pred' are numeric or integer}

\item{ylab}{label for the y-axis}

\item{xlab}{label for the y-axis}

\item{title}{title for the plot}
}
\value{
a ggplot2 object
}
\description{
Plot two-way partial dependence from random forests
}

