% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ci.R
\name{var_est.randomForest}
\alias{var_est.randomForest}
\title{Variance estimation for randomForest objects from package \code{randomForest}}
\usage{
\method{var_est}{randomForest}(fit, df)
}
\arguments{
\item{fit}{an object of class 'randomForest' returned from \code{randomForest} with \code{keep.inbag = TRUE}}

\item{df}{dataframe to be used for prediction}
}
\value{
a dataframe with two columns: 'prediction' and 'variance', where the former is the prediction calculated using the inbag data, and the variance is calculated using the bias corrected infinitesimal bootstrap from Wager, Efron, and Tibsharani (2014).
}
\description{
Calculates the variance of predictions from regression using randomForest by calling randomForestCI (\url{https://github.com/swager/randomForestCI})
}
\examples{
\dontrun{
library(randomForest)
data(swiss)

fit <- randomForest(Fertility ~ ., swiss, keep.inbag = TRUE)
var_est(fit, swiss)
}
}

