% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ci.R
\name{var_est.rfsrc}
\alias{var_est.rfsrc}
\title{Variance estimation for rfsrc objects from package \code{randomForestSRC}}
\usage{
\method{var_est}{rfsrc}(fit, df)
}
\arguments{
\item{fit}{an predict object of class 'rfsrc' returned from \code{rfsrc}}

\item{df}{dataframe to be used for prediction}
}
\value{
a dataframe with two columns: 'prediction' and 'variance', where the former is the prediction calculated using the inbag data, and the variance is calculated using the bias corrected infinitesimal bootstrap from Wager, Efron, and Tibsharani (2014).
}
\description{
Calculates the variance of predictions from regression using RandomForest using a slightly modified version of the code from randomForestCI (\url{https://github.com/swager/randomForestCI})
}
\examples{
\dontrun{
library(randomForestSRC)
data(swiss)

fit <- rfsrc(Fertility ~ ., swiss)
pred <- predict(fit, newdata = swiss)
fit$pd_membership <- pred$membership
fit$pd_predicted <- pred$predicted

var_est(fit, swiss)
}
}

