% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imp.R
\name{variable_importance.RandomForest}
\alias{variable_importance.RandomForest}
\title{Variable importance for RandomForest objects}
\usage{
\method{variable_importance}{RandomForest}(fit, conditional = FALSE,
  auc = FALSE, ...)
}
\arguments{
\item{fit}{an object of class 'randomForest' returned from \code{randomForest}}

\item{conditional}{logical, if true the conditional permutation importance is estimated, if not the marginal}

\item{...}{further arguments to be passed to varimp or varimpAUC}
}
\value{
a vector or dataframe of class "importance"
}
\description{
Extracts variable importances from a fitted \code{cforest}
}
\examples{
\dontrun{
library(party)
data(iris)

fit <- cforest(Species ~ ., iris, controls = cforest_control(mtry = 2))
variable_importance(fit)
}
}

