% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imp.R
\name{variable_importance}
\alias{variable_importance}
\title{Variable importance using random forests}
\usage{
variable_importance(fit, type, class_levels = FALSE, ...)
}
\arguments{
\item{fit}{object of class 'RandomForest', 'randomForest', or 'rfsrc'}

\item{type}{character, for randomForest equal to "accuracy", "gini", or "local"
if type is "accuracy" importance must be set to TRUE in the call to randomForest
for RandomForest (class for cforest) may be equal to "conditional" and/or "auc" (type may have length two), which will return the conditional/or auc importance}

\item{class_levels}{logical if possible return class specific importance, i.e. how much does the loss increase for each class from perturbing each feature in turn}

\item{...}{arguments to be passed to \code{variable_importance}}
}
\description{
Extracts variable importance for a set of predictors from a fitted random forest object from the party, randomForest, or randomForestSRC package
}
\examples{
\dontrun{
library(randomForest)
data(iris)

fit <- randomForest(Species ~ ., iris, importance = TRUE)
imp <- variable_importance(fit, "accuracy", TRUE)
plot_imp(imp)
}
}

