% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imp.R
\name{variable_importance.rfsrc}
\alias{variable_importance.rfsrc}
\title{Variable importance for rfsrc objects}
\usage{
\method{variable_importance}{rfsrc}(fit, type = "permute",
  class_levels = FALSE)
}
\arguments{
\item{fit}{an object of class 'rfsrc' returned from \code{rfsrc}}

\item{type}{character equal to "permute", "random", "permute.ensemble", or "random.ensemble"
this the \code{permute} argument must equal this value in the call to rfsrc}

\item{class_levels}{logical, when TRUE class level specific importances are returned otherwise the overal importance is returned}
}
\value{
a vector or dataframe of class "importance"
}
\description{
Extracts variable importances from a fitted \code{rfsrc}
}
\examples{
\dontrun{
library(randomForestSRC)
data(iris)

fit <- rfsrc(Species ~ ., iris, importance = "random")
variable_importance(fit, "random", TRUE)
}
}

