% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edarf.R
\docType{data}
\name{weed}
\alias{weed}
\title{Retail marijuana sales in Colorado}
\format{A data.frame with 2,807 observations and 8 variables:
\describe{
\itemize{
\item{city: name of nearest city}
\item{price: in U.S. dollars}
\item{amount: grams}
\item{quality: ordinal quality}
\item{date: date of purchase}
\item{ppg: price per gram}
\item{lat: latitude of city}
\item{lon: longitude of city}
}}}
\source{
\url{http://www.priceofweed.com}
}
\usage{
data(weed)
}
\description{
A dataset containing 2,807 retail marijuana transactions in Colorado from 2010-2013 from priceofweed.com
}
\keyword{datasets}

