% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prox.R
\name{extract_proximity.rfsrc}
\alias{extract_proximity.rfsrc}
\title{Extract proximity matrix from objects returned from \code{rfsrc}}
\usage{
\method{extract_proximity}{rfsrc}(fit)
}
\arguments{
\item{fit}{object of class 'rfsrc' from \code{rfsrc}}
}
\value{
an n by n matirx
}
\description{
Extract proximity matrix from objects returned from \code{rfsrc}
}

