% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pd.R
\name{partial_dependence}
\alias{partial_dependence}
\title{Partial dependence using random forests}
\usage{
partial_dependence(fit, df, var, cutoff = 10, interaction = FALSE,
  ci = TRUE, confidence = 0.95, empirical = TRUE, parallel = FALSE,
  type = "", ...)
}
\arguments{
\item{fit}{object of class 'RandomForest', 'randomForest', or 'rfsrc'}

\item{df}{the data.frame used to fit the model, only needed for randomForest}

\item{var}{a character vector of the predictors of interest, which must match the input matrix}

\item{cutoff}{the maximal number of unique points in each element of 'var' used in the
partial dependence calculation}

\item{interaction}{logical, if 'var' is a vector, does this specify an interaction or a list of bivariate partial dependence}

\item{ci}{use the bias corrected infinitesimal jackknife from Wager, Hastie, and Efron (2014), currently only works with regression}

\item{confidence}{desired confidence for the returned interval (ignored if ci is false)}

\item{empirical}{logical indicator of whether or not only values in the data should be sampled}

\item{parallel}{logical indicator of whether a parallel backend should be used if registered}

\item{type}{with classification, default "" gives most probable class for classification and "prob" gives class probabilities}

\item{...}{additional arguments to be passed}
}
\value{
a data.frame with the partial dependence of 'var'
if 'var' has length = 1 then the output will be a data.frame with a column for the predicted value at each value of 'var', averaged over the values of all other predictors.
if 'var' has length > 1 and interaction is false the out is returned in melted form, i.e. there is a column 'value''variable', and the predicted value for that combination.
if 'var' has length > 1 and interaction is true then the output will be a data.frame with a column for each element of 'var' and the predicted value for each combination.
}
\description{
Calculates the partial dependence of the response on an arbitrary dimensional set of predictors
from a fitted random forest object from the party, randomForest, or randomForestSRC packages
}
\examples{
\dontrun{
library(randomForest)
library(edarf)
## library(doParallel)
## library(parallel)
## registerDoParallel(makeCluster(detectCores()))

## Classification

data(iris)

fit <- randomForest(Species ~ ., iris)
pd <- partial_dependence(fit, iris, c("Petal.Width", "Sepal.Length"))
pd_int <- partial_dependence(fit, iris, c("Petal.Width", "Sepal.Length"), interaction = TRUE)

## Regression

data(swiss)

fit <- randomForest(Fertility ~ ., swiss, keep.inbag = TRUE)
pd <- partial_dependence(fit, swiss, "Education", ci = TRUE)
pd_int <- partial_dependence(fit, swiss, c("Education", "Catholic"), interaction = TRUE, ci = TRUE)
}
}

