% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pd}
\alias{plot.pd}
\title{Plot permutation importance from random forests}
\usage{
plot.pd(pd, title = NULL)
}
\arguments{
\item{pd}{object of class \code{c("pd", "data.frame")} as returned by
\code{\link{partial_dependence}}}

\item{title}{title for the plot}

\item{ylab}{label for the y-axis}

\item{xlab}{label for the y-axis}
}
\value{
a ggplot2 object
}
\description{
Plot permutation importance from random forests
}
\examples{
library(randomForest)
data(iris)
fit <- randomForest(Species ~ ., iris)
pd <- partial_dependence(fit, iris, "Petal.Width")
plot(pd)
}

