% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pd}
\alias{plot.pd}
\title{Plot partial dependence from random forests}
\usage{
\method{plot}{pd}(pd, geom = "line", title = NULL, facet_var = NULL)
}
\arguments{
\item{pd}{object of class \code{c("pd", "data.frame")} as returned by
\code{\link{partial_dependence}}}

\item{geom}{character describing type of plot desired: "bar", "line", or "area"}

\item{title}{title for the plot}

\item{facet_var}{A character vector indicating the variable that should be used
to facet on if inteaction is plotted. If not specified the variable with less
unique values is chosen.}
}
\value{
a ggplot2 object
}
\description{
Plot partial dependence from random forests
}
\examples{
library(randomForest)
library(edarf)
data(iris)
fit <- randomForest(Species ~ ., iris)
pd <- partial_dependence(fit, iris, "Petal.Width", type = "prob")
plot(pd, geom = "area")
}

