% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pd}
\alias{plot.pd}
\title{Plot permutation importance from random forests}
\usage{
plot.pd(pd, title = NULL, facet_var = NULL)
}
\arguments{
\item{pd}{object of class \code{c("pd", "data.frame")} as returned by
\code{\link{partial_dependence}}}

\item{title}{title for the plot}

\item{facet_var}{A caracter vector indicating the variable that should be used
to facet on if inteaction is plotted. If not specified the variable with less
unique values is chosen.}
}
\value{
a ggplot2 object
}
\description{
Plot permutation importance from random forests
}
\examples{
library(randomForest)
data(iris)
fit <- randomForest(Species ~ ., iris)
pd <- partial_dependence(fit, iris, "Petal.Width")
plot(pd)
}

