% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_imp}
\alias{plot_imp}
\title{Plot variable importance from random forests}
\usage{
plot_imp(imp, geom = "point", horizontal = TRUE, facet = FALSE,
  title = "")
}
\arguments{
\item{imp}{object of class \code{c("importance", "data.frame")} as returned by
\code{\link{variable_importance}}}

\item{geom}{character describing type of plot desired: "point" or "bar"}

\item{horizontal}{logical x-axis labels are horizontal if TRUE}

\item{facet}{logical indicating whether to facet, only applicable when returning class-specific variable importance}

\item{title}{title for the plot}
}
\value{
a ggplot2 object
}
\description{
Plot variable importance from random forests
}
\examples{
\dontrun{
library(randomForest)
data(iris)
fit <- randomForest(Species ~ ., iris, importance = TRUE)
imp <- variable_importance(fit, "accuracy", TRUE)
plot_imp(imp, "bar")
}
}

