% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_imp}
\alias{plot_imp}
\title{Plot variable importance from random forests}
\usage{
plot_imp(imp, sort = "none", labels = NULL, geom = "point",
  horizontal = TRUE, facet = FALSE, xlab = "Variables",
  ylab = "Importance", title = "")
}
\arguments{
\item{imp}{object of class \code{c("importance", "data.frame")} as returned by
\code{\link{variable_importance}}}

\item{sort}{character indicating if sorting of the output is to be done.
can be "none", "ascending", or "descending"}

\item{labels}{character vector giving labels to variables,
must have length equal to the number of rows in 'imp'}

\item{geom}{character describing type of plot desired: "point" or "bar"}

\item{horizontal}{logical x-axis labels are horizontal if TRUE}

\item{facet}{logical indicating whether to facet, only applicable when returning class-specific variable importance}

\item{xlab}{x-axis label, default "Variables"}

\item{ylab}{y-axis label, default "Importance"}

\item{title}{title for the plot}
}
\value{
a ggplot2 object
}
\description{
Plot variable importance from random forests
}
\examples{
\dontrun{
library(randomForest)
data(iris)
fit <- randomForest(Species ~ ., iris, importance = TRUE)
imp <- variable_importance(fit, "accuracy", TRUE)
plot_imp(imp, geom = "bar")
}
}

