% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_imp}
\alias{plot_imp}
\title{Plot permutation importance from random forests}
\usage{
plot_imp(var, imp, ylab = NULL, xlab = NULL, title = NULL)
}
\arguments{
\item{var}{character or factor vector of variable labels}

\item{imp}{numeric vector of variable permutation importance estimates}

\item{ylab}{label for the y-axis}

\item{xlab}{label for the y-axis}

\item{title}{title for the plot}
}
\value{
a ggplot2 object
}
\description{
Plot permutation importance from random forests
}

