% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prox}
\alias{plot_prox}
\title{Plot (sparse) principle components of the proximity matrix}
\usage{
plot_prox(prox, labels = NULL, size = 3, color = NULL,
  color_label = NULL, title = "", ...)
}
\arguments{
\item{prox}{n x n matrix giving the proportion of times across all trees that observation i,j are in the same terminal node}

\item{labels}{length n character vector giving observation labels}

\item{size}{integer which gives the size of points or text labels}

\item{color}{optional vector of length n which gives a factor with which to color the points}

\item{color_label}{character legend title if color parameter is used}

\item{title}{character plot title}

\item{...}{arguments to pass to \code{\link{prcomp}}}
}
\value{
a ggplot object
}
\description{
Plot (sparse) principle components of the proximity matrix
}
\examples{
\dontrun{
fit <- randomForest(hp ~ ., mtcars, proximity = TRUE)
prox <- extract_proximity(fit)
plot_prox(prox, labels = row.names(mtcars))

fit <- randomForest(Species ~ ., iris, proximity = TRUE)
prox <- extract_proximity(fit)
plot_prox(prox, color = iris$Species, color_label = "Species", size = 2)
}
}
\references{
https://github.com/vqv/ggbiplot

Gabriel, "The biplot graphic display of matrices with application to principal component analysis," \emph{Biometrika}, 1971
}

