% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ci.R
\name{var_est.RandomForest}
\alias{var_est.RandomForest}
\title{Variance estimation for RandomForest objects from package \code{party}}
\usage{
\method{var_est}{RandomForest}(fit, df)
}
\arguments{
\item{fit}{an object of class 'RandomForest' returned from \code{cforest}}

\item{df}{dataframe to be used for prediction}
}
\value{
a dataframe with two columns: 'prediction' and 'variance', where the former is the prediction calculated using the inbag data, and the variance is calculated using the bias corrected infinitesimal bootstrap from Wager, Efron, and Tibsharani (2014).
}
\description{
Calculates the variance of predictions from regression using RandomForest using a slightly modified version of the code from randomForestCI (\url{https://github.com/swager/randomForestCI})
}
\examples{
\dontrun{
library(party)
data(swiss)

fit <- cforest(Fertility ~ ., swiss, controls = cforest_control(mtry = 2))
var_est(fit, swiss)
}
}

