% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ci.R
\name{var_est}
\alias{var_est}
\title{Variance estimates for predicted values from random forests}
\usage{
var_est(fit, df)
}
\arguments{
\item{fit}{object of class 'RandomForest', 'randomForest', or 'rfsrc' (must be regression)}

\item{df}{dataframe to be used for prediction}
}
\description{
Calculates the variance of predictions from a random forest using the bias corrected infinitesimal
jackknife from Wager, Efron, and Tibsharani (2014) using a fitted random forest object from the
party, randomForest, or randomForestSRC packages
}

