% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{var_est}
\alias{var_est}
\title{Variance estimates for predicted values from random forests}
\usage{
var_est(fit, data, calibrate, ...)
}
\arguments{
\item{fit}{object of class 'RandomForest', 'randomForest', or 'rfsrc' (must be regression)}

\item{data}{dataframe to be used for prediction}

\item{calibrate}{whether to calibrate the variance estimates using empirical bayes via randomForestCI, default is TRUE}

\item{...}{additional arguments to be passed to the predict method}
}
\value{
a data.frame of length n with one column 'prediction' which contains the ensemble prediction
and a column 'variance' which contains the estimated variance
}
\description{
Calculates the variance of predictions from a random forest using the bias corrected infinitesimal
jackknife from Wager, Efron, and Tibsharani (2014) using a fitted random forest object from the
party, randomForest, or randomForestSRC packages
}
\references{
Wager, Hastie, and Efron, "Confidence intervals for random forests: the jackknife and the infinitesimal jackknife," \emph{Journal of Machine Learning Research}, 2014.

https://github.com/swager/randomForestCI
}

