% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imp.R
\name{variable_importance}
\alias{variable_importance}
\title{Variable importance using random forests}
\usage{
variable_importance(fit, var, type, oob, nperm, parallel, ...)
}
\arguments{
\item{fit}{object of class 'RandomForest', 'randomForest', or 'rfsrc'}

\item{var}{character, variables to find the importance of}

\item{type}{character, either "aggregate," (default) which gives the average loss under permutation or "local" which gives the residual across all observations}

\item{nperm}{positive integer giving the number of times to permute the indicated variables (default 10)}

\item{parallel}{logical whether to run in parallel using the registered backend (default FALSE)}

\item{...}{additional arguments to be passed to \code{variable_importance}. the randomForest method requires an additional argument \code{data} which is the training data.frame}
}
\description{
Computes variable importance for a set of predictors from a fitted random forest object from the party, randomForest, or randomForestSRC package
}
\examples{
\dontrun{
library(randomForest)
data(iris)

fit <- randomForest(Species ~ ., iris, importance = TRUE)
variable_importance(fit, var = colnames(iris)[-5], type = "aggregate", nperm = 10, data = iris)

library(party)
data(swiss)
fit <- cforest(Fertility ~ ., swiss)
variable_importance(fit, var = colnames(swiss)[-1], type = "local", nperm = 10)

data(mtcars)
library(randomForestSRC)
fit <- rfsrc(mpg ~ ., mtcars)
variable_importance(fit, var = colnames(mtcars)[-1], type = "aggregate", nperm = 10, oob = TRUE)
}
}

