% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imp.R
\name{variable_importance}
\alias{variable_importance}
\title{Variable importance using random forests}
\usage{
variable_importance(fit, ...)
}
\arguments{
\item{fit}{object of class 'RandomForest', 'randomForest', or 'rfsrc'}

\item{...}{arguments to be passed to \code{variable_importance}}
}
\description{
Extracts variable importance for a set of predictors from a fitted random forest object from the party, randomForest, or randomForestSRC package
}

